Extract Metriks::Report
