+++ lib/metriks/reporter/graphite.rb
@@ -1,9 +1,10 @@
 require 'socket'
+require 'metriks/reporter'
 require 'metriks/registry'
 require 'metriks/time_tracker'
 
-module Metriks::Reporter
-  class Graphite
+class Metriks::Reporter
+  class Graphite < Metriks::Reporter
     attr_accessor :host, :port, :prefix
 
     def initialize(host, port, options = {})
@@ -11,40 +12,12 @@ module Metriks::Reporter
       @port   = port
       @prefix = options[:prefix]
 
-      @registry     = options[:registry] || Metriks::Registry.default
-      @time_tracker = Metriks::TimeTracker.new(options[:interval] || 60)
-      @on_error     = options[:on_error] || proc { |ex| }
+      super options
     end
 
     def socket
       @socket = nil if @socket && @socket.closed?
       @socket ||= TCPSocket.new(@host, @port)
-    end
-
-    def start
-      @thread ||= Thread.new do
-        loop do
-          @time_tracker.sleep
-
-          Thread.new do
-            begin
-              write
-            rescue Exception => ex
-              @on_error[ex] rescue nil
-            end
-          end
-        end
-      end
-    end
-
-    def stop
-      @thread.kill if @thread
-      @thread = nil
-    end
-
-    def restart
-      stop
-      start
     end
 
     def write
