+++ lib/metriks/reporter/logger.rb
@@ -1,9 +1,10 @@
 require 'logger'
+require 'metriks/reporter'
 require 'metriks/registry'
 require 'metriks/time_tracker'
 
-module Metriks::Reporter
-  class Logger
+class Metriks::Reporter
+  class Logger < Metriks::Reporter
     attr_accessor :prefix, :log_level, :logger
 
     def initialize(options = {})
@@ -11,33 +12,7 @@ module Metriks::Reporter
       @logger    = options[:logger]    || ::Logger.new(STDOUT)
       @prefix    = options[:prefix]    || 'metriks:'
 
-      @registry     = options[:registry] || Metriks::Registry.default
-      @time_tracker = Metriks::TimeTracker.new(options[:interval] || 60)
-      @on_error     = options[:on_error] || proc { |ex| }
-    end
-
-    def start
-      @thread ||= Thread.new do
-        loop do
-          @time_tracker.sleep
-
-          begin
-            write
-          rescue Exception => ex
-            @on_error[ex] rescue nil
-          end
-        end
-      end
-    end
-
-    def stop
-      @thread.kill if @thread
-      @thread = nil
-    end
-
-    def restart
-      stop
-      start
+      super options
     end
 
     def flush
