+++ test/logger_reporter_isolated_test.rb
@@ -94,81 +94,7 @@ class LoggerReporterIsolatedTest < Test::Unit::TestCase
     reporter.write
   end
 
-  def test_specify_registry
-    registry = stub
-    registry.expects(:each)
-    reporter = Metriks::Reporter::Logger.new(:registry => registry)
-    reporter.write
-  end
-
-  def test_uses_default_registry
-    registry = stub
-    registry.expects(:each)
-    Metriks::Registry.expects(:default).returns(registry)
-    Metriks::Reporter::Logger.new.write
-  end
-
-  # TODO: These tests don't really test the interval. Tie them in with #start
-  # tests when written.
-  def test_specify_interval
-    Metriks::TimeTracker.expects(:new).with(42)
-    Metriks::Reporter::Logger.new(:interval => 42)
-  end
-
-  def test_default_interval_one_minute
-    Metriks::TimeTracker.expects(:new).with(60)
-    Metriks::Reporter::Logger.new
-  end
-
-  def test_specify_error_handler
-    Metriks::TimeTracker.any_instance.stubs(:sleep)
-    handler = stub
-    handler.expects(:[]).with(kind_of(RuntimeError)).at_least_once
-    reporter = Metriks::Reporter::Logger.new(:on_error => handler)
-    def reporter.write() raise('write') end
-
-    reporter.start
-    sleep 0.01
-    reporter.stop
-  end
-
-  def test_swallows_errors_raised_in_error_handler
-    Metriks::TimeTracker.any_instance.stubs(:sleep)
-    handler = stub
-    handler.expects(:[]).raises(RuntimeError).at_least_once
-    reporter = Metriks::Reporter::Logger.new(:on_error => handler)
-    def reporter.write() raise('write') end
-
-    reporter.start
-    sleep 0.01
-    reporter.stop
-  end
-
-  def test_default_error_handler_swallows_errors
-    Metriks::TimeTracker.any_instance.stubs(:sleep)
-    reporter = Metriks::Reporter::Logger.new
-    def reporter.write() raise('write') end
-
-    reporter.start
-    sleep 0.01
-    reporter.stop
-  end
-
   ### Public Methods
-
-  def test_stop
-    Metriks::TimeTracker.any_instance.stubs(:sleep)
-    reporter = Metriks::Reporter::Logger.new
-    reporter.start
-    reporter.stop
-    reporter.expects(:write).never
-    sleep 0.01
-  end
-
-  def test_stop_unstarted_reporter
-    reporter = Metriks::Reporter::Logger.new
-    assert_nothing_raised { reporter.stop }
-  end
 
   def test_write_includes_time
     Time.stubs(:now => 42)
