+++ test/proc_title_reporter_isolated_test.rb
@@ -50,76 +50,7 @@ class ProcTitleReporterIsolatedTest < Test::Unit::TestCase
     assert_equal "#{ @original_proctitle } counter: 1", $0
   end
 
-  # TODO: These tests don't really test the interval. Tie them in with #start
-  # tests when written.
-  def test_specify_interval
-    Metriks::TimeTracker.expects(:new).with(42)
-    Metriks::Reporter::ProcTitle.new(:interval => 42)
-  end
-
-  def test_default_interval_one_minute
-    Metriks::TimeTracker.expects(:new).with(60)
-    Metriks::Reporter::ProcTitle.new
-  end
-
-  def test_specify_error_handler
-    Metriks::TimeTracker.any_instance.stubs(:sleep)
-    handler = stub
-    handler.expects(:[]).with(kind_of(RuntimeError)).at_least_once
-    reporter = Metriks::Reporter::ProcTitle.new(:on_error => handler)
-    def reporter.write() raise('write') end
-
-    reporter.start
-    sleep 0.01
-    reporter.stop
-  end
-
-  def test_swallows_errors_raised_in_error_handler
-    Metriks::TimeTracker.any_instance.stubs(:sleep)
-    handler = stub do
-      expects(:[]).raises(RuntimeError).at_least_once
-    end
-    reporter = Metriks::Reporter::ProcTitle.new(:on_error => handler)
-    def reporter.write() raise('write') end
-
-    reporter.start
-    sleep 0.01
-    reporter.stop
-  end
-
-  def test_default_error_handler_swallows_errors
-    Metriks::TimeTracker.any_instance.stubs(:sleep)
-    reporter = Metriks::Reporter::ProcTitle.new
-    def reporter.write() raise('write') end
-
-    reporter.start
-    sleep 0.01
-    reporter.stop
-  end
-
   ### Public Methods
-
-  def test_stop
-    Metriks::TimeTracker.any_instance.stubs(:sleep)
-    reporter = Metriks::Reporter::ProcTitle.new
-    reporter.start
-    sleep 0.01
-    reporter.stop
-    # Let some threads spawn and wait for them to die.
-    wait = 0
-    while Thread.list.size > 0
-      wait += 1
-      break if wait > 10
-      sleep 0.01
-    end
-    reporter.expects(:write).never
-    sleep 0.01
-  end
-
-  def test_stop_unstarted_reporter
-    reporter = Metriks::Reporter::ProcTitle.new
-    assert_nothing_raised { reporter.stop }
-  end
 
   def test_empty_with_no_metrics
     reporter = Metriks::Reporter::ProcTitle.new
